+++ packages/core/navigation-next/docs/examples/ui-components/Item.js
@@ -4,9 +4,10 @@ import React from 'react';
 import Badge from '@atlaskit/badge';
 import CloseButton from '@atlaskit/icon/glyph/cross';
 import { Item, ItemAvatar } from '../../../src';
+import { CONTENT_NAV_WIDTH } from '../../../src/common/constants';
 
 export default () => (
-  <div css={{ width: 270 }}>
+  <div css={{ width: CONTENT_NAV_WIDTH }}>
     <Item
       after={({ isActive, isHover }) =>
         isActive || isHover ? (