+++ packages/editor/editor-core/src/__tests__/integration/card/inline-1.ts
@@ -9,7 +9,6 @@ import {
   clipboardInput,
 } from '../_helpers';
 
-// --adding comments for testing
 [fullpage].forEach(editor => {
   BrowserTestCase(
     `pasting an link converts to inline card`,