+++ packages/editor/editor-core/src/plugins/collab-edit/provider/channel.ts
@@ -7,6 +7,7 @@ import {
   PubSubClient,
   DocumentResponse,
   StepResponse,
+  MixedResponse,
 } from './types';
 import { logger } from './';
 
@@ -107,7 +108,7 @@ export class Channel {
   /**
    * Send steps to service
    */
-  async sendSteps(state: any, getState: () => any) {
+  async sendSteps(state: any, getState: () => any, localSteps?: any[]) {
     if (this.isSending) {
       this.debounce(getState);
       return;
@@ -120,7 +121,7 @@ export class Channel {
       return;
     }
 
-    const { steps = [] } = sendableSteps(state) || {}; // sendableSteps can return null..
+    const { steps = [] } = localSteps || sendableSteps(state) || {}; // sendableSteps can return null..
 
     if (steps.length === 0) {
       logger(`No steps to send. Aborting.`);
@@ -155,21 +156,12 @@ export class Channel {
    * Get steps from version x to latest
    */
   async getSteps(version: number) {
-    try {
-      const response = await utils.requestService<DocumentResponse>(
-        this.config,
-        {
-          path: `document/${this.config.docId}/steps`,
-          queryParams: {
-            version,
-          },
-        },
-      );
-
-      this.emit('data', response);
-    } catch (err) {
-      logger(`Unable to get latest steps: ${err}`);
-    }
+    return await utils.requestService<MixedResponse>(this.config, {
+      path: `document/${this.config.docId}/steps`,
+      queryParams: {
+        version,
+      },
+    });
   }
 
   /**