+++ packages/editor/editor-core/src/plugins/table/pm-plugins/table-resizing/index.ts
@@ -1,0 +1,2 @@
+export { columnResizing, key } from './plugin';
+export { scaleTable, setColumnWidths } from './actions';